Functions

Functions

Defining variables

Logical operations

Arithmetic operations

Comparison operators

Generating the SMT representation of a problem

Solving a SAT problem

Miscellaneous functions

Test link link

Defining variables

Core.Bool — Method

Bool("z")

Construct a single Boolean variable with name "z".

    Bool(n, "z")
    Bool(m, n, "z")

Construct a vector-valued or matrix-valued Boolean variable with name "z".

Vector and matrix-valued Booleans use Julia's built-in array functionality: calling Bool(n,"z") returns a Vector{BoolExpr}, while calling Bool(m, n, "z") returns a Matrix{BoolExpr}.

Core.Int — Method

Int("a")

Construct a single Int variable with name "a".

    Int(n, "a")
    Int(m, n, "a")

Construct a vector-valued or matrix-valued Int variable with name "a".

Vector and matrix-valued Ints use Julia's built-in array functionality: calling Int(n,"a") returns a Vector{IntExpr}, while calling Int(m, n, "a") returns a Matrix{IntExpr}.

Core.Real — Method

Real("r")

Construct a single Int variable with name "r".

    Real(n, "r")
    Real(m, n, "r")

Construct a vector-valued or matrix-valued Real variable with name "r".

Vector and matrix-valued Reals use Julia's built-in array functionality: calling Real(n,"a") returns a Vector{RealExpr}, while calling Real(m, n, "r") returns a Matrix{RealExpr}.

Logical operations

These are operations in the theory of propositional logic. For a formal definition of this theory, see Figure 3.2 in The SMT-LIB Standard, Version 2.6.

BooleanSatisfiability.not — Method

not(z::BoolExpr)
¬z

Return the logical negation of z.

Note: Broacasting a unary operator requires the syntax .¬z which can be confusing to new Julia users. We define ¬(z::Array{BoolExpr}) for convenience.

    z = Bool(n, "z")
    ¬z  # syntactic sugar for map(¬, z)
    .¬z # also valid

BooleanSatisfiability.and — Method

z1 ∧ z2
and(z1,...,zn)
and([z1,...,zn])

Returns the logical AND of two or more variables. Use dot broadcasting for vector-valued and matrix-valued Boolean expressions.

z1 = Bool(n, "z1")
z2 = Bool(m, n, "z2")
z1 .∧ z2
and.(z1, z2) # equivalent to z1 .∧ z2

Special cases:

and(z) returns z.

and(z, false) returns false.

and(z, true) returns z.

BooleanSatisfiability.or — Method

z1 ∨ z2
or(z1,...,zn)
or([z1,...,zn])

Returns the logical OR of two or more variables. Use dot broadcasting for vector-valued and matrix-valued Boolean expressions.

z1 = Bool(n, "z1")
z2 = Bool(m, n, "z2")
z1 .∨ z2
or.(z1, z2) # equivalent to z1 .∨ z2

Special cases:

or(z) returns z.

or(z, false) returns z.

or(z, true) returns true.

Note that ∨ (\vee) is NOT the ASCII character v.

Base.xor — Method

xor(z1,...,zn)
⊻(z1,...zn)

XOR (exclusive or) is true if exactly one of z1,...,zn is true and false otherwise. Use dot broadcasting across arrays.

Special cases:

xor(z) returns z.

xor(false, z) returns z.

xor(true, z) returns ¬z.

xor(true, true, z) returns false.

BooleanSatisfiability.implies — Method

z1 ⟹ z2
implies(z1, z2)

Returns the expression z1 IMPLIES z2. Use dot broadcasting for vector-valued and matrix-valued Boolean expressions. Note: implies(z1, z2) is equivalent to or(not(z1), z2).

BooleanSatisfiability.iff — Method

iff(z1::BoolExpr, z2::BoolExpr)
z1 ⟺ z2

Bidirectional implication between z1 and z2. Equivalent to and(z1 ⟹ z2, z2 ⟹ z1).

BooleanSatisfiability.ite — Method

ite(x::BoolExpr, y::BoolExpr, z::BoolExpr)

If-then-else statement. Equivalent to or(x ∧ y, ¬x ∧ z).

Base.all — Method

all([z1,...,zn])

Return and(z1,...,zn). If z1,...,zn are themselves AND operations, all(z)flattens the nestedAND`.

Examples:

and([and(z1, z2), and(z3, z4)]) == and(z1, z2, z3, z4)

and([or(z1, z3), z3, z4]) == and(or(z1, z3), z3, z4)

Base.any — Method

any([z1,...,zn])

Return or(z1,...,zn). If z1,...,zn are themselves OR operations, any(z)flattens the nestedOR`. Examples:

any([or(z1, z2), or(z3, z4)]) == or(z1, z2, z3, z4)

any([and(z1, z3), z3, z4]) == or(and(z1, z3), z3, z4)

Arithmetic operations

These are operations in the theory of integer and real-valued arithmetic.

Note that +, -, and * follow type promotion rules: if both a and b are IntExprs, a+b will have type IntExpr. If either a or b is a RealExpr, the result will have type RealExpr. Division \ is defined only in the theory of real-valued arithmetic, thus it always has return type RealExpr. For a formal definition of the theory of integer arithmetic, see Figure 3.3 in The SMT-LIB Standard, Version 2.6.

Base.:- — Method

-(a::IntExpr)
-(r::RealExpr)

Return the negative of an Int or Real expression.

    -Int(n, "a")# this works
    -Int(n, m, "b") # this also works

Base.:+ — Method

a + b
a + 1 + true

Return the Int | Real expression a+b (inherits the type of a+b). Use dot broadcasting for vector-valued and matrix-valued Boolean expressions.

a = Int(n, "a")
b = Int(n, m, "b")
a .+ b
println("typeof a+b: $(typeof(a[1] + b[1]))")
c = Real("c")
println("typeof a+c: $(typeof(a[1] + c))")
z = Bool("z")
a .+ z
println("typeof a+z: $(typeof(a[1] + z))")

Base.:- — Method

a - b
a - 2

Returns the Int | Real expression a-b (inherits the type of a-b). Use dot broadcasting for vector-valued and matrix-valued Boolean expressions.

a = Int(n, "a")
b = Int(n, m, "b")
a .- b
println("typeof a-b: $(typeof(a[1] - b[1]))")
c = Real("c")
println("typeof a-c: $(typeof(a[1] - c))")
z = Bool("z")
a .- z
println("typeof a-z: $(typeof(a[1] - z))")

Base.:* — Method

a * b
a * 2

Returns the Int | Real multiplication expression a*b (inherits the type of a*b). Use dot broadcasting for vector-valued and matrix-valued Boolean expressions.

a = Int(n, "a")
b = Int(n, m, "b")
a .* b
println("typeof a*b: $(typeof(a[1]*b[1]))")
c = Real("c")
println("typeof a*c: $(typeof(a[1]*c))")
z = Bool("z")
a .- z
println("typeof a*z: $(typeof(a[1]*z))")

Base.:/ — Method

a / b
a / 1.0

Returns the Real division expression a/b. Note: a and b must be Real). Use dot broadcasting for vector-valued and matrix-valued Boolean expressions.

a = Real(n, "a")
b = Real(n, m, "b")
a ./ b
println("typeof a/b: $(typeof(a[1]/b[1]))")

Comparison operators

Missing docstring.

Missing docstring for Base.:(==)(a::AbstractExpr, b::AbstractExpr). Check Documenter's build log for details.

Base.:< — Method

a < b
a < 0

Returns the Boolean expression a < b. Use dot broadcasting for vector-valued and matrix-valued expressions.

a = Int(n, "a")
b = Int(n, m, "b")
a .< b
z = Bool("z")
a .< z

Base.:<= — Method

a <= b
a <= 0

Returns the Boolean expression a <= b. Use dot broadcasting for vector-valued and matrix-valued expressions.

a = Int(n, "a")
b = Int(n, m, "b")
a .<= b
z = Bool("z")
a .<= z

Base.:> — Method

a > b
a > 0

Returns the Boolean expression a > b. Use dot broadcasting for vector-valued and matrix-valued expressions.

a = Int(n, "a")
b = Int(n, m, "b")
a .> b
z = Bool("z")
a .> z

Base.:>= — Method

a >= b
a >= 0

Returns the Boolean expression a >= b. Use dot broadcasting for vector-valued and matrix-valued expressions.

a = Int(n, "a")
b = Int(n, m, "b")
a .>= b
z = Bool("z")
a .>= z

Generating the SMT representation of a problem

BooleanSatisfiability.smt — Method

smt(z::AbstractExpr)
smt(z1,...,zn)
smt([z1,...,zn])

Generate the SMT representation of z or and(z1,...,zn).

When calling smt([z1,...,zn]), the array must have type Array{AbstractExpr}. Note that list comprehensions do not preserve array typing. For example, if z is an array of BoolExpr, [z[i] for i=1:n] will be an array of type Any. To preserve the correct type, use BoolExpr[z[i] for i=1:n].

BooleanSatisfiability.save — Method

save(z::AbstractExpr, filename)
save(z::Array{AbstractExpr}, filename=filename)
save(z1, z2,..., filename)                  # z1, z2,... are type AbstractExpr

Write the SMT representation of z or and(z1,...,zn) to filename.smt.

Solving a SAT problem

BooleanSatisfiability.sat! — Method

sat!(z::BoolExpr)
sat!(z1, z2,...)

Solve the SAT problem using Z3. If the problem is satisfiable, update the values of all BoolExprs in prob with their satisfying assignments.

Possible return values are :SAT, :UNSAT, or :ERROR. prob is only modified to add Boolean values if the return value is :SAT.

BooleanSatisfiability.value — Method

value(z::BoolExpr)
value(z::Array{BoolExpr})

Returns the satisfying assignment of z, or nothing if no satisfying assignment is known. In the array-valued case, returns Array{Bool} or Array{nothing}.

It's possible to return an array of mixed Bool and nothing. This could occur if some variables in an array do not appear in a problem, because sat!(problem) will not set the values of variables that do not appear in problem.

Miscellaneous functions

Base.isequal — Method

Test equality of two AbstractExprs.